% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amort_schedule.R
\name{amort_schedule}
\alias{amort_schedule}
\title{Amortization Schedule}
\usage{
amort_schedule(principal, rate, years)
}
\arguments{
\item{principal}{Single numeric value - The total loan amount.}

\item{rate}{Single numeric value - The annual interest rate (as a decimal).}

\item{years}{Single numeric value - The total number of years for the loan.}
}
\value{
A data frame showing the breakdown of each payment.
}
\description{
This function performs a calculation with robust input validation.
}
\examples{

# With user provided variables:

# Rate can be entered as a whole number (conversion handled by check in function)
# or as a decimal. The function will determine the rate to be a percentage
# between 0 and 1

testP<-1000
testR<-5
testY<-10
amort_schedule(testP,testR,testY)

}
