% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fv_simple.R
\name{fv_simple}
\alias{fv_simple}
\title{Simple Future Value}
\usage{
fv_simple(principal, rate, years)
}
\arguments{
\item{principal}{Single numeric value - Must not contain \code{NA} values and all values must be positive.}

\item{rate}{Single numeric percentage - Must not contain \code{NA} values and all values must be positive.}

\item{years}{Single numeric value - Must not contain \code{NA} values and all values must be positive.}
}
\value{
A numeric value representing the result of the calculation. The future value.
}
\description{
This function performs a calculation with robust input validation.
}
\examples{

# Using user variables:

testP<-1000
testR<-5
testY<-10
fv_simple(testP,testR,testY)

}
