% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loan_payment.R
\name{loan_payment}
\alias{loan_payment}
\title{Loan Payment}
\usage{
loan_payment(principal, rate, years)
}
\arguments{
\item{principal}{Single numeric value - The total loan amount.}

\item{rate}{Single numeric value - The annual interest rate (as a decimal).}

\item{years}{Single numeric value - The total number of years for the loan.}
}
\value{
A numeric value representing the result of the calculation. The fixed monthly payment.
}
\description{
This function performs a calculation with robust input validation.
}
\examples{

# Using user variables:

testP<-1000
testR<-5
testY<-10
loan_payment(testP,testR,testY)

}
