% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles1.R
\docType{data}
\name{profiles1}
\alias{profiles1}
\title{Data Set: 'profiles1' - 100 records for financial calculations:}
\format{
\subsection{\code{profiles1}}{

A data frame with 100 rows and 9 columns:
\describe{
\item{Principal}{Initial amount invested}
\item{Rate}{Annual interest rate (as a decimal)}
\item{Years}{Duration of investment in years}
\item{N_Componding}{Number of compounding periods per year}
\item{Future_Value_Input}{Value at a future date}
\item{Start_Value}{Initial value of an investment}
\item{End_Value}{Final value of an investment}
\item{Start_Price}{Initial price of a stock}
\item{End_Price}{Final price of a stock}
}
}
}
\source{
\url{https://github.com/rwalshUSF/SimpleFinCalc/blob/main/data/profiles1.rda}
}
\usage{
profiles1
}
\description{
synthetic inputs for various loan, investment, and return calculations
for use with the functions implemented in the 'SimpleFinCalc' package.
}
\keyword{datasets}
