% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv.R
\name{pv}
\alias{pv}
\title{Present Value}
\usage{
pv(future_value, rate, years)
}
\arguments{
\item{future_value}{Single numeric value - The future value amount.}

\item{rate}{Single numeric value - The annual discount rate (as a decimal).}

\item{years}{Single numeric value - The number of years until the future value is received.}
}
\value{
A numeric value representing the result of the calculation. The present value.
}
\description{
This function performs a calculation with robust input validation.
}
\examples{

# With user provided variables:

# Rate can be entered as a whole number (conversion handled by check in function)
# or as a decimal. The function will determine the rate to be a percentage
# between 0 and 1

testF<-10000
testR<-5
testY<-10
pv(testF,testR,testY)

}
