% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annuity_payment.R
\name{annuity_payment}
\alias{annuity_payment}
\title{Annuity Payments}
\usage{
annuity_payment(principal, rate, years)
}
\arguments{
\item{principal}{Single numeric value - The initial principal amount (loan amount or present value).}

\item{rate}{Single numeric value - The annual interest rate (as a decimal).}

\item{years}{Single numeric value - The number of years for the annuity/loan.}
}
\value{
A numeric value representing the result of the calculation. The regular payment amount.
}
\description{
This function performs a calculation with robust input validation.
}
\examples{

# Using user variables:

testP<-1000
testR<-5
testY<-10
annuity_payment(testP,testR,testY)

}
