% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cagr.R
\name{cagr}
\alias{cagr}
\title{Compound Annual Growth Rate}
\usage{
cagr(start_value, end_value, years)
}
\arguments{
\item{start_value}{Single numeric value - The initial value of the investment.}

\item{end_value}{Single numeric value - The final value of the investment.}

\item{years}{Single numeric value - The number of years the investment was held.}
}
\value{
The Compound Annual Growth Rate (CAGR) of an investment as a decimal.
}
\description{
This function performs a calculation with robust input validation.
}
\examples{

# Using user variables:

testSV<-1000
testEV<-15632
testY<-10
fv_simple(testSV,testEV,testY)

}
