% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fv_compound.R
\name{fv_compound}
\alias{fv_compound}
\title{Compound Future Value}
\usage{
fv_compound(principal, rate, years, n_comp)
}
\arguments{
\item{principal}{The initial principal amount (numeric, positive).}

\item{rate}{The annual interest rate (numeric, positive, as a decimal).}

\item{years}{The number of years (numeric, positive integer or decimal).}

\item{n_comp}{The number of compounding periods per year (numeric, positive integer).}
}
\value{
The calculated future value (numeric). The future value.
}
\description{
This function performs a calculation with robust input validation.
}
\examples{

# With user provided variables:

# Rate can be entered as a whole number (conversion handled by check in function)
# or as a decimal. The function will determine the rate to be a percentage
# between 0 and 1

testP<-1000
testR<-5
testY<-10
testN<-4
fv_compound(testP,testR,testY,testN)

}
