% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fv_simple.R
\name{fv_simple}
\alias{fv_simple}
\title{Simple Future Value}
\usage{
fv_simple(principal, rate, years)
}
\arguments{
\item{principal}{Single numeric value - The initial principal amount (numeric, positive).}

\item{rate}{Single numeric percentage - The annual interest rate (numeric, positive, as a decimal).}

\item{years}{Single numeric value - The number of years (numeric, positive integer or decimal).}
}
\value{
A numeric value representing the result of the calculation. The future value.
}
\description{
This function performs a calculation with robust input validation.
}
\examples{

# With user provided variables:

# Rate can be entered as a whole number (conversion handled by check in function)
# or as a decimal. The function will determine the rate to be a percentage
# between 0 and 1

testP<-1000
testR<-5
testY<-10
fv_simple(testP,testR,testY)

}
