% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fv_simple.R
\name{fv_simple}
\alias{fv_simple}
\title{Simple Future Value}
\usage{
fv_simple(principal, rate, years)
}
\arguments{
\item{principal}{Single numeric value}

\item{rate}{Single numeric percentage}

\item{years}{Single numeric value}
}
\value{
A numeric value
}
\description{
Simple Future Value
}
\examples{
testP<-1000
testR<-5
testY<-10
fv_simple(testP,testR,testY)

}
